backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy[postgresql]==2.0.31
alembic==1.13.2

# Limiter
slowapi==0.1.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic[email]==2.7.4

# AI / NLP
spacy==3.7.5
sentence-transformers==3.0.1
httpx==0.27.0

# File processing
PyMuPDF==1.24.5
python-docx==1.1.2
pdfplumber==0.11.1

# Utilities
loguru==0.7.2
python-dotenv==1.0.1
